+++ resources/views/katalog.blade.php
@@ -68,31 +68,45 @@
                         </div>
                 </div>
             </div>
-            <div class="row">
-                <div class="col-xs-6 col-sm-1 text-left text-sm-right year-range">
-                        <span class="sans" id="from_year">{!! !empty($input['year-range']) ? reset((explode(',', $input['year-range']))) : App\Item::sliderMin() !!}</span>
+            <div class="row mt-10">
+                <div class="col-xs-6 col-sm-1 text-left text-sm-right">
+                    <input class="sans" id="from_year" maxlength="4" pattern="[0-9]{1-4}" step="5" value="{!! !empty($input['year-range']) ? reset((explode(',', $input['year-range']))) : App\Item::sliderMin() !!}" />
                 </div>
-                <div class="col-xs-6 col-sm-1 col-sm-push-10 text-right text-sm-left year-range">
-                        <span class="sans" id="until_year">{!! !empty($input['year-range']) ? end((explode(',', $input['year-range']))) : App\Item::sliderMax() !!}</span>
+                <div class="col-xs-6 col-sm-1 col-sm-push-10 text-right text-sm-left ">
+                    <input class="sans" id="until_year"  maxlength="4" pattern="[0-9]{1-4}" step="5" value="{!! !empty($input['year-range']) ? end((explode(',', $input['year-range']))) : App\Item::sliderMax() !!}" />
                 </div>
-                <div class="col-sm-10 col-sm-pull-1 year-range">
-                        <input id="year-range" name="year-range" type="text" class="span2" data-slider-min="{!! App\Item::sliderMin() !!}" data-slider-max="{!! App\Item::sliderMax() !!}" data-slider-step="5" data-slider-value="[{!! !empty($input['year-range']) ? $input['year-range'] : App\Item::sliderMin().','.App\Item::sliderMax() !!}]"/>
+                <div class="col-xs-12 col-sm-10 col-sm-pull-1">
+                    @include('components.year_slider', ['id' => 'yearRangeFilter'])
+                    @include('components.year_slider_js', [
+                        'yearRange' => !empty($input['year-range']) ? $input['year-range'] : App\Item::sliderMin() . ', ' . App\Item::sliderMax(),
+                        'min' => App\Item::sliderMin(), 
+                        'max' => App\Item::sliderMax(), 
+                        'id' => 'yearRangeFilter'
+                        ])
                 </div>
             </div>
-            @if ($color)
             <div class="row">
-                <div class="col-sm-12">
-                    <label for="color_filter" class="w-100 mt-3 mb-0 light">
-                        {{ utrans('katalog.filters_color') }}:
-                        @include('components.color_list', ['colors' => [array('hex' => '#'.$color, 'amount' => '100%')], 'include_clear' => true, 'id' => 'color-filter', 'class_names' => 'mt-2 mb-0'])
-                        {!! Form::hidden('color', @$input['color'], ['id'=>'color']) !!}
+                <div class="col-sm-12 col-md-10 col-md-push-1">
+                    <div class="color-picker">
+                        @include('components.color_picker', ['id'=>'colorpicker'])
+                        @include('components.color_picker_js', ['id' => 'colorpicker', 'color' => $color])
+                    </div>
+                </div>
+
+                @if ($color)
+                <div class="col-sm-12 col-md-10 col-md-push-1">
+                    <label for="color_filter" class="w-100 b-0 light">
+                        @include('components.color_list', ['colors' => [array('hex' => '#'.$color, 'amount' => '100%')], 'include_clear' => true, 'id' => 'color-filter', 'class_names' => 'mb-0'])
                     </label>
                 </div>
+                @endif
             </div>
-            @endif
+            {!! Form::hidden('color', @$input['color'], ['id'=>'color']) !!}
             {!! Form::hidden('sort_by', @$input['sort_by'], ['id'=>'sort_by']) !!}
+            {!! Form::hidden('year-range', @$input['year-range'], ['id'=>'year-range']) !!}
             {!! Form::close() !!}
-    </div></div>
+        </div>
+    </div>
 </section>
 
 @foreach ($items as $i=>$item)
@@ -210,7 +224,6 @@ $('.isotope-wrapper').each(function(){
 });
 
 $(document).ready(function(){
-
     // $('.expandable').readmore({
     //     moreLink: '<a href="#" class="text-center">viac možností <i class="icon-arrow-down"></i></a>',
     //     lessLink: '<a href="#" class="text-center">menej možností <i class="icon-arrow-up"></i></a>',
@@ -239,16 +252,31 @@ $(document).ready(function(){
         }
     });
 
-    $("#year-range").slider({
-        // value: [1800, 1900],
-        tooltip: 'hide'
-    }).on('slideStop', function(event) {
-        $(this).closest('form').submit();
-    }).on('slide', function(event) {
-        var rozsah = $("#year-range").val().split(',');
-        $('#from_year').html(rozsah[0]);
-        $('#until_year').html(rozsah[1]);
+    yearRangeFilter.$on('change', function(range) {
+        $('#year-range').val(range.join(','));
+        $('#filter').submit();
     });
+    yearRangeFilter.$on('slide', function(range) {
+        $('#from_year').val(range[0]);
+        $('#until_year').val(range[1]);
+    });
+
+    colorpicker.$on('change', function(clr){
+        $('#color').val(clr.hex.substr(1));
+        $('#filter').submit();
+    })
+    $('#from_year,#until_year').on('change', function(event){
+        const min = {!! App\Item::sliderMin() !!};
+        const max = {!! App\Item::sliderMax() !!};
+        const fy = +$('#from_year').val().replace(/\D/g, '')
+        const uy = +$('#until_year').val().replace(/\D/g, '');
+        const from = Math.min(Math.max(min, fy), max);
+        const until = Math.max(Math.min(max, uy), min);
+        $('#year-range').val([from,until].sort().join(','));
+        $('#filter').submit();
+    })
+
+  
 
     // $(".custom-select").chosen({allow_single_deselect: true})
     $(".custom-select").selectize({
@@ -270,7 +298,7 @@ $(document).ready(function(){
         }
     });
 
-    $(".custom-select, input[type='checkbox']").change(function() {
+    $(".custom-select, input[type='checkbox']:not(#use_color)").change(function() {
         var form = $(this).closest('form');
         form.submit();
     });